fix(alert): radio and checkbox labels wrap to next line (#27898)
Issue number: resolves #17269

---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

Long radio and checkbox labels truncate with ellipsis instead of
wrapping to the next line. This makes the label hard to understand
because users cannot read the entire label.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Labels now wrap to the next line instead of truncating.

Note: Some of the screenshots may show wider alerts than before. This is
the correct behavior. When rendering text, the browser will try to
render as much text on a single line as it can. Since our alerts can
expand in width, the alert wrapper will expand in width (up to the max
width) to accommodate this text. Once the the wrapper is at the max
width, text will then wrap to the next line.

For example, you may notice the MD alert increasing to 280px in width:
https://github.com/ionic-team/ionic-framework/blob/b8553d89f84970db452646ad7a88e2fc0fd88657/core/src/components/alert/alert.md.vars.scss#L11.

There should be no visual diff for alerts with text that would not
normally wrap to the next line.

This was not happening before because we did not allow text to wrap to
the next line.

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this introduces a breaking change, please describe the impact
and migration path for existing applications below. -->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

---------

Co-authored-by: ionitron <hi@ionicframework.com>